fix: adicion de manejo generico de excepciones ante insersiones y modificaciones, y conversion de metodo findRoutesNotSell a JPA derived methods

+++ src/main/java/unlp/info/bd2/repositories/ReviewRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import unlp.info.bd2.model.Review;
 
 @Repository
-public interface ReviewRepository extends CrudRepository<Review, Integer> {
+public interface ReviewRepository extends CrudRepository<Review, Long> {
 
 }
